requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
pkg-resources==0.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
